octoxlabs/requirements.txt
asgiref==3.7.2
attrs==23.2.0
black==24.2.0
cfgv==3.4.0
click==8.1.7
coverage==7.4.3
distlib==0.3.8
Django==5.0.3
django-stubs==4.2.7
django-stubs-ext==4.2.7
djangorestframework==3.14.0
drf-spectacular==0.27.1
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==24.1.0
filelock==3.13.1
flake8==7.0.0
identify==2.5.35
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mccabe==0.7.0
mypy==1.9.0
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.6.2
pycodestyle==2.11.1
pyflakes==3.2.0
pytest==8.1.1
pytest-cov==4.1.0
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.18.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
types-pytz==2024.1.0.20240203
types-PyYAML==6.0.12.20240311
typing_extensions==4.10.0
uritemplate==4.1.1
virtualenv==20.25.1
